+++ cfolk/Final_Project/app.py
@@ -6,7 +6,9 @@ from flask_mongoengine import MongoEngine
 
 app = Flask(__name__)
 
+# app.config["MONGO_URI"] = "mongodb://localhost:27017/six_months_db"
 app.config["MONGO_URI"] = "mongodb://localhost:27017/stock_db"
+
 mongo = PyMongo(app)
 
 @app.route('/')
@@ -16,7 +18,7 @@ def index():
 
 @app.route('/all')
 def read_all():
-    users = mongo.db.top_ten.find()
+    users = mongo.db.six_months.find()
     output = {'All': []}
     # cycle through users
     for user in users:
