+++ cfolk/Final_Project/static/plot.js
@@ -27,7 +27,7 @@ function handleSubmit() {
 
       for (i=0; i < info.length; i++) {
         var name = info[i].symbol;
-        
+        // console.log(info);
         // if stock ticker entered = stock ticker in json
         if (stock === name) {
 
@@ -37,19 +37,20 @@ function handleSubmit() {
           var closingPrices = info[i].historical.map(row=>row['close']);
           // console.log(closingPrices);
 
-
-          // converting historical dates/prices for predictive chart
-          var dateLength = dates.length;
-          // console.log(dateLength);
-          var predictChartDates = dates.slice(dateLength - 365);
-          // console.log(predictChartDates);
-          var priceLength = closingPrices.length;
-          var predictChartPries = closingPrices.slice(priceLength - 365);
-
           // variables from predictive array
           var predictionDates = info[i].prediction[0].prediction_data['Date'];
           var predictionPrice = info[i].prediction[0].prediction_data['Predictions'];
           // console.log(predictionDates);
+          // console.log(predictionPrice);
+          // var predictionDatesClean = predictionDates[]
+
+          // // converting historical dates/prices for predictive chart
+          var dateLength = predictionDates.length;
+          // console.log(dateLength);
+          var predictChartDates = predictionDates.slice(dateLength - 180);
+          // console.log(predictChartDates);
+          var priceLength = predictionPrice.length;
+          var predictChartPries = predictionPrice.slice(priceLength - 180);
 
         }
       }
@@ -109,16 +110,16 @@ function handleSubmit() {
         }
       }
 
-      // var trace3 = {
-      //   type: "scatter",
-      //   mode: "lines",
-      //   name: "Actual",
-      //   x: predictChartDates,
-      //   y: predictChartPries,
-      //   line: {
-      //     color: "#17BECF"
-      //   }
-      // };
+      var trace3 = {
+        type: "scatter",
+        mode: "lines",
+        name: "Actual",
+        x: predictChartDates,
+        y: predictChartPries,
+        line: {
+          color: "#17BECF"
+        }
+      };
 
       var data1 = [trace1];
 
@@ -135,10 +136,10 @@ function handleSubmit() {
         showlegend: true    
       };
 
-      var data2 = [trace2, trace1];
+      var data2 = [trace2, trace3];
 
       var layout2 = {
-        title: `${stock} Predictive Chart`,
+        title: `${stock} Predictive Chart Last 60 Days`,
         xaxis: {
           range: [startPredictFormat, endDate], //endPredictFormat
           type: "date"
