+++ cfolk/Final_Project/templates/index.html
@@ -20,11 +20,11 @@
         <div class="title">
           <h1 class="avengers"><strong>The Avengers</strong></h1>
           <h2 class="avengers"><strong>Earths Mightiest Investors</strong></h2>
-          <h5 class="subheader">Top 10 most popular stocks of 2020: </h5>
-          <h5 class="options">AAPL, AMZN, AMD, FB, MSFT, MRNA, NIO, NKLA, NVDA, TSLA</h5> 
-          <!-- <p class="hover_txt">::HOVER OVER ME::</p> -->
+          <!-- <h5 class="subheader">Top 10 most popular stocks of 2020: </h5>
+          <h5 class="options">AAPL, AMZN, AMD, FB, MSFT, MRNA, NIO, NKLA, NVDA, TSLA</h5>  -->
+          <p class="hover_txt">::HOVER OVER ME::</p>
           <form class="d-flex">
-            <input class="form-control me-2" id= "stockInput" name = "stock" type="search" placeholder="Enter One of the Above Stock Tickers" aria-label="Search">
+            <input class="form-control me-2" id= "stockInput" name = "stock" type="search" placeholder="Enter Nasdeq Stock Ticker" aria-label="Search">
             <button class="btn btn-outline-success" id="submit" type="submit">Search</button>
           </form>
         </div>
